requirements.txt
entsoe-py==0.7.8
pandas==2.3.3
numpy==2.3.5
sqlalchemy==2.0.45
dotenv==0.9.9
plotly==6.5.0
nbformat==5.10.4
scikit-learn==1.3.0
xgboost==2.0.0
holidays==0.35
matplotlib==3.10.8
optuna==3.0.0
pytest==8.0.0